Change type of completion.CreateParams.Prompt
The OpenAI API accepts a single string, or an array of strings. When
representing as a Go struct, we need to choose one or the other.

A slice of strings is more general, but if we only ever send singleton
lists in practice, then actually a single string is more convenient.